Pagination method added

+++ Models/Customer.cs
@@ -7,7 +7,7 @@ namespace mr_mat_api.Models
         public string City { get; set; }
         public string State { get; set; }
         public string Country { get; set; }
-        public int PinCode { get; set; }
+        public long PinCode { get; set; }
         public string Mobile { get; set; }
         public string Email { get; set; }
         public string Gender { get; set; }
